Add `@echo off` to relevant functions
When a funtion is called using `call` it inherits the @echo state
from its parent, when it is called using `start` it resets it to default.
The relevant functions are functions that are often used in
asynchronous processes, in which they are called using `start`. This
makes the child-processes output a ton of data which makes it
difficult to spot output or debug any errors.