Implement functions/echo.bat
`echo.bat` is a unified function for displaying text in the console.
it can output text in colors to make console output more readable
also, the :verbose function checks for the value of %verbose% to
provide output only if it is needed.

Make sure to surround the message string with quotes.